now the bubbles is a components
