+++ frontend/src/App.css
@@ -96,16 +96,17 @@ background-attachment: fixed;
 }
 
 .composent-button-link {
-      width: 8rem;
+      width: 8.5rem;
       height: 4rem;
       text-align: center;
       align-items: flex-end;
       box-shadow: inset 4px 8px 8px rgba(0, 0, 0, 0.25);
-      border-radius: 15%;
-      color: purple;
-      background-color: rgb(168, 50, 189);
+      border-radius: 20px 20px 20px 20px;
+      color: rgb(210, 151, 210);
+      background-color: #9a48d0;
       margin-top: 15%;
       z-index: 10;
+      padding: 4px;
 }
 
 a.composent-button-link {
@@ -172,8 +173,10 @@ a.composent-button-link {
 
 .floatting-text {
   font-family: var(--text-main-font);
-  font-size: small;
+  font-size:  0.9rem;
   justify-content: center;
   display: flex;
+  margin-top: 8%;
 }
+
 /*Les titres et textes*/
