+++ frontend/src/pages/Home.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable import/no-unresolved */
+/* eslint-disable import/extensions */
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-plusplus */
 import React from "react";
@@ -5,54 +7,13 @@ import Logos from "../components/Logos";
 import Button from "../components/Button";
 import FloattingText from "../components/FloattingText";
 import FloattingTitle from "../components/FloattingTitle";
-
-// Les couleurs des bulles
-const colors = ["#54F98D", "#9683EC", "#E73E01", "#DF73FF"];
-
-// La création des bulles
-const numBalls = 60;
-const balls = [];
-
-for (let i = 0; i < numBalls; i++) {
-  const ball = document.createElement("div");
-  ball.classList.add("ball");
-  ball.style.background = colors[Math.floor(Math.random() * colors.length)];
-  ball.style.left = `${Math.floor(Math.random() * 100)}vw`;
-  ball.style.top = `${Math.floor(Math.random() * 100)}vh`;
-  ball.style.transform = `scale(${Math.random()})`;
-  ball.style.width = `${Math.random()}em`;
-  ball.style.height = ball.style.width;
-
-  balls.push(ball);
-  document.body.append(ball);
-}
-
-// Le mouvement des bulles
-balls.forEach((el, i) => {
-  const to = {
-    x: Math.random() * (i % 2 === 0 ? -11 : 11),
-    y: Math.random() * 12,
-  };
-
-  const anim = el.animate(
-    [
-      { transform: "translate(0, 0)" },
-      { transform: `translate(${to.x}rem, ${to.y}rem)` },
-    ],
-    {
-      duration: (Math.random() + 1) * 2000, // random duration
-      direction: "alternate",
-      fill: "both",
-      iterations: Infinity,
-      easing: "ease-in-out",
-    }
-  );
-});
+import LittleBubbles from "../components/LittleBubbles";
 
 export default function Home() {
   return (
     <div className="home">
       <Logos />
+      <LittleBubbles />
       <FloattingTitle title="Adopte un reste, c'est quoi ?" />
       <FloattingText
         text="Cette année plus que jamais, la générosité et la solidarité sont de mise, alors à l'action !
